+++ src/components/dropdown/index.tsx
@@ -12,6 +12,7 @@ const DropdownMenu = ({
   isDisabled,
 }: DropDownMenuType) => {
   const handleChange = (newSortType: MovieSortOptions) => {
+    console.log('1. in DropdownMenu, in handleChange, good');
     onSortChange(newSortType);
   };
