+++ src/pages/homepage/index.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useFullMovies } from '../../pages/homepage/hooks';
 import { MovieSortOptions } from '../../types';
 import Header from '../../components/header';
-import MovieList from '../../components/movieList/MovieList2';
+import MovieList from '../../components/movieList';
 import Navbar from '../../components/navbar/Navbar';
 import DropdownMenu from '../../components/dropdown';
 
@@ -25,8 +25,8 @@ const Homepage = () => {
       <Navbar />
       <DropdownMenu
         sortType={sortType}
-        onSortChange={() => setSortType(sortType)}
-        onResetMovies={() => setSortType('newest')}
+        onSortChange={handleSortChange}
+        onResetMovies={handleResetMovies}
         isDisabled={isDisabled}
       />
       <MovieList sortType={sortType} />
